* commit '826bc990f56698ddf021c11e09ed8d220da7c993':
  [BL-369] Undo using apply instead of commit (shared preferences) as commit was explicitly used.
  [BL-369] Undo using apply instead of commit (shared preferences) as commit was explicitly used.
  [BL-369] Removed default template and update/add copyright headers
  [BL-369] Update/add copyright headers
  [BL-369] More specific errors & messages
  [BL-369] More locale fixes
  [BL-369] Unneeded casting
  [BL-369] Unused imports (really??)
  [BL-369] Use apply instead of commit (shared preferences)
  [BL-369] Use US locale for dates (default if not specified)
  [BL-369] Invert comparisons to get correct errors
  [BL-369] Remove assert lint warnings

+++ Source/ShineLib/shinelib/src/test/java/com/philips/pins/shinelib/utility/DeviceInformationCacheTest.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) Koninklijke Philips N.V., 2015, 2016.
+ * All rights reserved.
+ */
+
 package com.philips.pins.shinelib.utility;
 
 import com.philips.pins.shinelib.capabilities.SHNCapabilityDeviceInformation;
@@ -31,7 +36,6 @@ public class DeviceInformationCacheTest {
     @Before
     public void setUp() throws ParseException {
         initMocks(this);
-
 
         deviceInformationCache = new DeviceInformationCache(persistentStorageMock);
     }
@@ -69,7 +73,7 @@ public class DeviceInformationCacheTest {
                 boolean res = false;
 
                 if (argument instanceof Long) {
-                    long millis = ((Long) argument).longValue();
+                    long millis = (Long) argument;
                     min = TEST_DATE.getTime();
                     max = new Date().getTime();
